Add new function to copy images with autocompletion feature
Added a new zsh function 'yamano:cpimg' in 'work.zsh' to securely copy images from a remote host, with a pem key for authentication. Also introduced a new autocompletion method 'yamano:cpimg_comp' to offer host suggestions based on entries in the 'yamano_config' file. This reduces the risk of typos and increases efficiency when working with remote image files. The new script 'work.zsh' is sourced in 'package.zsh' to load on zsh startup.