Update Brewfile to include 'parallels' cask
This commit has updated the 'brew/.Brewfile' by adding the 'parallels' cask, which was absent in the previous version of the file. This inclusion simplifies the process of installing Parallels software with the help of Homebrew while setting up the system using the Brewfile. The rest of the file remains unchanged.